+++ CHANGELOG.md
@@ -2,15 +2,17 @@
 
 ## 0.4.0
 
-+ Added `ValueProperty`, abstract superclass for `ConfigProperty`.
++ Added `ValueProperty`, abstract superclass for `ConfigProperty`
++ Changed the behavior of `makelist()`: now it can also decorate a callable, converting its return type to a list
++ Added `addattr()`
 
 ## 0.3.6
 
-- Fixed `ConfigValue` behavior with multiple sources. It used to iterate through all the sources, possibly overwriting; now, iteration stops at first non-missing value.
+- Fixed `ConfigValue` behavior with multiple sources. It used to iterate through all the sources, possibly overwriting; now, iteration stops at first non-missing value
 
 ## 0.3.5
 
-- Fixed cb32 handling. Now leading zeros in SIQ's are stripped, and `.from_cb32()` was implemented.
+- Fixed cb32 handling. Now leading zeros in SIQ's are stripped, and `.from_cb32()` was implemented
 
 ## 0.3.4