+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## 0.4.0
 
 + Added `ValueProperty`, abstract superclass for `ConfigProperty`
-+ Changed the behavior of `makelist()`: now it can also decorate a callable, converting its return type to a list
++ \[BREAKING] Changed the behavior of `makelist()`: now it's also a decorator, converting its return type to a list (revertable with `wrap=False`)
++ New module `lex` with functions `symbol_table()` and `lex()` — make tokenization more affordable
 + Added `addattr()`
 
 ## 0.3.6